requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.11.8
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
astor==0.8.1
async-timeout==4.0.3
asyncio-atexit==1.0.1
attrs==24.2.0
backoff==2.2.1
bce-python-sdk==0.9.23
beautifulsoup4==4.12.3
bidict==0.23.1
boto3==1.35.78
botocore==1.35.78
bottle==0.13.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
chinese_calendar==1.10.0
click==8.1.7
coloredlogs==15.0.1
colorlog==6.9.0
cryptography==44.0.0
cssselect==1.2.0
cssutils==2.11.1
cycler==0.12.1
dataclasses-json==0.6.7
decorator==5.1.1
deepdiff==8.0.1
distro==1.9.0
docutils==0.21.2
emoji==2.14.0
erniebot==0.5.0
erniebot_agent==0.5.0
et_xmlfile==2.0.0
einops==0.8.1
exceptiongroup==1.2.2
faiss-cpu==1.8.0.post1
fastapi==0.115.8
filelock==3.13.1
filetype==1.2.0
flatbuffers==24.3.25
fonttools==4.55.0
frozenlist==1.5.0
fsspec==2024.2.0
ftfy==6.3.1
future==1.0.0
gdown==5.2.0
GPUtil==1.4.0
greenlet==3.1.1
h11==0.14.0
html5lib==1.1
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
huggingface-hub==0.26.3
humanfriendly==10.0
idna==3.10
ifaddr==0.2.0
imageio==2.36.1
imagesize==1.4.1
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.0
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-path==0.3.3
jsonschema-specifications==2023.12.1
kiwisolver==1.4.7
langchain==0.1.5
langchain-community==0.0.17
langchain-core==0.1.23
langdetect==1.0.9
langsmith==0.0.87
lazy-object-proxy==1.10.0
lazy_loader==0.4
lxml==5.3.0
markdown2==2.5.3
MarkupSafe==2.1.5
marshmallow==3.23.1
matplotlib==3.5.2
more-itertools==10.5.0
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.2.1
nicegui==2.10.1
nltk==3.9.1
numpy==1.24.4
olefile==0.47
onnxruntime==1.20.1
openai==1.55.3
openapi-schema-validator==0.6.2
openapi-spec-validator==0.7.1
opencv-python==4.10.0.84
openpyxl==3.1.5
opt-einsum==3.3.0
orderly-set==5.2.2
orjson==3.10.15
packaging==23.2
paddlepaddle==3.0.0b2
paddlex @ https://paddle-model-ecology.bj.bcebos.com/paddlex/whl/paddlex-3.0.0b2-py3-none-any.whl
pandas==2.2.3
Parsley==1.3
pathable==0.4.3
pdftext==0.3.19
pillow==10.4.0
pluggy==1.5.0
premailer==3.10.0
prettytable==3.12.0
propcache==0.2.0
protobuf==5.28.3
proxy-tools==0.1.0
pscript==0.7.7
psutil==6.1.0
pycairo==1.27.0
pyclipper==1.3.0.post6
pycocotools==2.0.8
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.2
pydantic-settings==2.6.1
pydantic_core==2.27.1
Pygments==2.19.1
PyGObject==3.50.0
PyMuPDF==1.24.14
pyparsing==3.2.0
pypdf==5.1.0
pypdfium2==4.30.0
PyQt6==6.8.0
PyQt6-Qt6==6.8.1
PyQt6-WebEngine==6.8.0
PyQt6-WebEngine-Qt6==6.8.1
PyQt6_sip==13.10.0
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.11.2
python-iso639==2024.10.22
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.1
python-socketio==5.12.1
pytz==2024.2
pywebview==5.4
PyYAML==6.0.2
qianfan==0.0.3
QtPy==2.4.2
RapidFuzz==3.10.1
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rpds-py==0.21.0
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.12
s3transfer==0.10.4
safetensors==0.4.5
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
shapely==2.0.6
simple-websocket==1.1.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
starlette==0.45.3
surya-ocr==0.6.13
sympy==1.13.1
tabled-pdf==0.1.4
tabulate==0.9.0
tenacity==8.5.0
threadpoolctl==3.5.0
tifffile==2024.9.20
tokenizers==0.20.3
tomli==2.2.1
torch @ https://download.pytorch.org/whl/cpu/torch-2.5.1%2Bcpu-cp310-cp310-linux_x86_64.whl
tqdm==4.67.1
transformers==4.46.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
ujson==5.10.0
unstructured==0.16.8
unstructured-client==0.25.9
urllib3==2.2.3
uvicorn==0.34.0
uvloop==0.21.0
vbuild==0.8.2
watchfiles==1.0.4
wcwidth==0.2.13
webencodings==0.5.1
websockets==14.2
wrapt==1.17.0
wsproto==1.2.0
yarl==1.18.0
